Comment route verified through middleware
