+++ server/package.json
@@ -8,7 +8,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "seed": "node src/seed.js",
     "start": "node index.js",
-    "dev": "nodemon index.js""
+    "dev": "nodemon index.js"
   },
   "keywords": [
     "technobyte",
