+++ server/src/router/routes.js
@@ -4,6 +4,7 @@ import { adminMiddleware } from "../middleware/admin.middleware.js";
 import postAdmin from "./post-admin.js";
 import { reviewBlog } from "../controllers/blog-review.js";
 import newsRoutes from "./newsletter.js";
+import { userMiddleware } from "../middleware/user.middleware.js";
 
 import commentroute from "./comment.js"
 function routes(app) {
@@ -13,7 +14,7 @@ function routes(app) {
 
     // app.use('/auth', authRoutes);
     app.use("/newsletter", newsRoutes);
-    app.use("blogs/comment",commentroute);
+    app.use("blogs/comment",userMiddleware,commentroute);
     app.post("/admin", adminMiddleware, postAdmin);
     app.use("/admin/blog/review/:id", reviewBlog);
 }
